+++ trusttunnel/setup_wizard/src/template_settings.rs
@@ -37,6 +37,9 @@ vpn_mode = "{}"
 killswitch_enabled = {}
 
 {}
+killswitch_allow_ports = []
+
+{}
 post_quantum_group_enabled = {}
 
 {}
@@ -51,6 +54,7 @@ dns_upstreams = []
     Settings::default_vpn_mode(),
     Settings::doc_killswitch_enabled().to_toml_comment(),
     Settings::default_killswitch_enabled(),
+    Settings::doc_killswitch_allow_ports().to_toml_comment(),
     Settings::doc_post_quantum_group_enabled().to_toml_comment(),
     Settings::default_post_quantum_group_enabled(),
     Settings::doc_exclusions().to_toml_comment(),