+++ trusttunnel/src/client.cpp
@@ -392,6 +392,7 @@ VpnListener *TrustTunnelClient::make_tun_listener(ListenerSettings listener_sett
     VpnWinTunnelSettings win_settings = *vpn_win_tunnel_settings_defaults();
     win_settings.wintun_lib = m_wintun;
     win_settings.block_untunneled = m_config.killswitch_enabled;
+    win_settings.block_untunneled_exclude_ports = m_config.killswitch_allow_ports.c_str();
     VpnError res = m_tunnel->init(&tunnel_settings, &win_settings);
 #else
 # ifdef __linux__